requirements.txt
Flask==2.0.2
pytube==11.0.1
moviepy==1.0.3
youtube-search-python==1.4.5
